move mock adapter and fixture to conftest
